Add backup thread to backup db every x hours.

+++ sanji/model_initiator.py
@@ -6,14 +6,16 @@ import os
 import shutil
 import subprocess
 import time
-import threading
+from threading import Thread
+from threading import Event
+from threading import RLock
 
 
 class ModelInitiator(object):
     """
     "   Deal with some model initialization works like DB
     "   and Condifuration files creating.
-    "   [backup_inteval]: backup db every hours(s)
+    "   [backup_inteval]: backup db every hours(s), minus means no backup.
     """
     def __init__(
             self, model_name, model_path, db_type="json", backup_interval=720):
@@ -29,11 +31,13 @@ class ModelInitiator(object):
             self.model_name + ".json"
         self.db_type = db_type
         self.backup_interval = backup_interval * 3600  # hour
-        self.db_mutex = threading.RLock()
+        self.db_mutex = RLock()
         self.db_manager()
 
+        self._backup_thread = None
+        self._backup_thread_event = Event()
         if self.backup_interval > 0:
-            self.periodic_backup_db()
+            self._backup_thread = self.periodic_backup_db()
 
     def db_manager(self):
         """
@@ -44,7 +48,10 @@ class ModelInitiator(object):
             self.load_db()
         except Exception as e:
             print "***", e
-            self.recover_db(self.backup_json_db_path)
+            try:
+                self.recover_db(self.backup_json_db_path)
+            except Exception:
+                pass
         else:
             return True
 
@@ -77,14 +84,22 @@ class ModelInitiator(object):
         " [src_file]: copy from src_file to xxxxx.json
         """
         with self.db_mutex:
-            shutil.copy2(src_file, self.json_db_path)
+            try:
+                shutil.copy2(src_file, self.json_db_path)
+            except IOError as e:
+                print "*** No %s file." % src_file
+                raise e
 
     def backup_db(self):
         """
         " Generate a xxxxx.backup.json.
         """
         with self.db_mutex:
-            shutil.copy2(self.json_db_path, self.backup_json_db_path)
+            if os.path.exists(self.json_db_path):
+                try:
+                    shutil.copy2(self.json_db_path, self.backup_json_db_path)
+                except OSError:
+                    print "*** No file to copy."
 
     def load_db(self):
         """
@@ -110,18 +125,39 @@ class ModelInitiator(object):
             else:
                 self.sync()
 
+    def start_backup(self):
+        if self._backup_thread.is_alive():
+            raise RuntimeError("Stop previous backup thread first.")
+
+        self._backup_thread.start()
+
+    def stop_backup(self):
+        if self._backup_thread:
+            if self._backup_thread.is_alive():
+                self._backup_thread_event.set()
+                self._backup_thread.join()
+                return True
+        return False
+
     def periodic_backup_db(self):
         """
         " Fork a thread to backup db periodically.
         """
-        t = threading.Thread(target=self.thread_backup_db)
+        t = Thread(target=self.thread_backup_db)
         t.daemon = True
         t.start()
+        return t
 
     def thread_backup_db(self):
-        while True:
-            self.backup_db()
-            time.sleep(self.backup_interval)
+        single_sleep_time = 3
+        sleep_count = self.backup_interval
+        while not self._backup_thread_event.is_set():
+            if sleep_count >= self.backup_interval:
+                self.backup_db()
+                sleep_count = 0
+            else:
+                time.sleep(single_sleep_time)
+                sleep_count += single_sleep_time
 
     def sync(self):
         """
@@ -129,6 +165,3 @@ class ModelInitiator(object):
         """
         cmd = "sync"
         subprocess.call(cmd, shell=True)
-
-    def __del__(self):
-        pass
